Removed unused stuff
~ Changed: the filenames that where written in PascalCase to be in camelCase. Affected imports have been adjusted.
Reason: to be consistent with the rest of the code, and the Google Styleguide (note: not specified in GTS, but done in all typescript project by Google on GitHub).

~ Changed: The class User to declare the types of the variables instead of making them class fields.
Reason: Sequelize documentation (see: https://sequelize.org/docs/v6/core-concepts/model-basics/) states that public class fields with the same name as one of the model's attributes will override the setters/getters created by running init on a model. Using 'declare' makes it so typescript doesn't complain about types, while not shadowing the setters/getters. (Note: this hasn't created problems yet, but could in the future).

- Removed: commented out code from userDAO.ts, userDTO.ts, user.ts.
Reason: they were used before we used used the old database, but not anymore.

- Removed: 'getTransactionMgr()' and 'createTables()' from database.ts.
Reason: Throughout th code we access the sequelize object 'database' directly, so 'getTransactionMgr' is unnecessary. We currently assume tables in sequelize models match the database and therefore don't sync. Because of this we previously created a separate method for authenticating with the database, making 'createTables()' redundant.